+++ changelog.md
@@ -5,7 +5,13 @@
 * Villagers Attacking
   * No longer attack enemies (can be configured)
   * No longer attack players with high enough reputation (defaults to Iron Golem)
+* Spiders Throwing Webs
+  * Cave Spiders now also apply poison when hitting entities (configurable)
+  * Added config option to place a cobweb on the entity hit by the thrown web
+  * Added config option to disable slowness effect correctly
+  * Reduced max slowness level (~~VI (6)~~ -> IV (4))
 * Animal knockback can now be configured
+* Drowned now go in Swimming pose when swimming, quite ugly as they don't support the swimming animation fully
 
 ## Beta 1.7.1
 * Requires InsaneLib 1.7.1